refactor: remove web chat-thread service leftovers
