--- turbo/apps/web/src/lib/zero/chat-thread/chat-thread-service.ts
@@ -1,28 +1,0 @@
-import { eq } from "drizzle-orm";
-import { chatThreads } from "@vm0/db/schema/chat-thread";
-import { publishThreadListChanged } from "./chat-message-service";
-
-/**
- * Update a chat thread's title.
- */
-export async function updateChatThreadTitle(
-  threadId: string,
-  userId: string,
-  title: string,
-): Promise<void> {
-  const [thread] = await globalThis.services.db
-    .select({ renamedAt: chatThreads.renamedAt })
-    .from(chatThreads)
-    .where(eq(chatThreads.id, threadId))
-    .limit(1);
-
-  if (thread?.renamedAt) {
-    return;
-  }
-
-  await globalThis.services.db
-    .update(chatThreads)
-    .set({ title })
-    .where(eq(chatThreads.id, threadId));
-  await publishThreadListChanged(userId);
-}
