--- turbo/apps/web/src/lib/zero/chat-thread/index.ts
@@ -1,1 +1,0 @@
-export { updateChatThreadTitle } from "./chat-thread-service";
